Rain: moved the resolution of int (to a 64-bit signed integer) out to a pass

+++ RainParse.hs
@@ -177,7 +177,7 @@ name
 dataType :: RainParser A.Type
 dataType 
   = do {sBool ; return A.Bool}
-    <|> do {sInt ; return A.Int64}
+    <|> do {sInt ; return A.Int}
     <|> do {sChannel ; inner <- dataType ; return $ A.Chan A.DirUnknown (A.ChanAttributes {A.caWritingShared = False, A.caReadingShared = False}) inner}
     <|> do {sIn ; inner <- dataType ; return $ A.Chan A.DirInput (A.ChanAttributes {A.caWritingShared = False, A.caReadingShared = False}) inner}
     <|> do {sOut ; inner <- dataType ; return $ A.Chan A.DirOutput (A.ChanAttributes {A.caWritingShared = False, A.caReadingShared = False}) inner}
